+++ firestore-bigquery-export/firestore-bigquery-change-tracker/src/__tests__/bigquery/partitioning.test.ts
@@ -6,6 +6,7 @@ import { ChangeType, FirestoreDocumentChangeEvent } from "../..";
 import { FirestoreBigQueryEventHistoryTrackerConfig } from "../../bigquery";
 import { Partitioning } from "../../bigquery/partitioning";
 import { deleteTable } from "../fixtures/clearTables";
+import { changeTracker } from "../fixtures/changeTracker";
 
 let bq: BigQuery;
 let dataset: Dataset;
@@ -197,6 +198,110 @@ describe("processing partitions on a new table", () => {
       expect(value.end_date).toBeDefined();
     });
 
+    test("returns a value when timePartitioningField and timePartitioningFirestoreField string value has been defined, with a timestamp-like value", async () => {
+      const config: FirestoreBigQueryEventHistoryTrackerConfig = {
+        datasetId: "",
+        tableId: "",
+        datasetLocation: "",
+        timePartitioning: "",
+        timePartitioningField: "end_date",
+        timePartitioningFieldType: "DATETIME",
+        timePartitioningFirestoreField: "end_date",
+        transformFunction: "",
+        clustering: [],
+        bqProjectId: null,
+      };
+
+      // a Timestamp-Like object (we lose the instance after serialization)
+      const end_date = {
+        _seconds: 1614153600,
+        _nanoseconds: 0,
+      };
+
+      const event: FirestoreDocumentChangeEvent = {
+        timestamp: "",
+        operation: ChangeType.CREATE,
+        documentName: "",
+        eventId: "",
+        documentId: "",
+        data: { end_date },
+      };
+
+      const partitioning = new Partitioning(config, table);
+      const value = partitioning.getPartitionValue(event);
+
+      expect(value.end_date).toBeDefined();
+    });
+
+    test("returns an empty object when _seconds or _nanoseconds is not a number", async () => {
+      const config: FirestoreBigQueryEventHistoryTrackerConfig = {
+        datasetId: "",
+        tableId: "",
+        datasetLocation: "",
+        timePartitioning: "",
+        timePartitioningField: "end_date",
+        timePartitioningFieldType: "DATETIME",
+        timePartitioningFirestoreField: "end_date",
+        transformFunction: "",
+        clustering: [],
+        bqProjectId: null,
+      };
+
+      // a Timestamp-Like object (we lose the instance after serialization)
+      const end_date = {
+        _seconds: "not a number",
+        _nanoseconds: 0,
+      };
+
+      const event: FirestoreDocumentChangeEvent = {
+        timestamp: "",
+        operation: ChangeType.CREATE,
+        documentName: "",
+        eventId: "",
+        documentId: "",
+        data: { end_date },
+      };
+
+      const partitioning = new Partitioning(config, table);
+      const value = partitioning.getPartitionValue(event);
+
+      expect(value).toEqual({});
+    });
+
+    test("returns a value when timePartitioningField and timePartitioningFirestoreField string value has been defined, and is timestamp-like", async () => {
+      const config: FirestoreBigQueryEventHistoryTrackerConfig = {
+        datasetId: "",
+        tableId: "",
+        datasetLocation: "",
+        timePartitioning: "",
+        timePartitioningField: "end_date",
+        timePartitioningFieldType: "DATETIME",
+        timePartitioningFirestoreField: "end_date",
+        transformFunction: "",
+        clustering: [],
+        bqProjectId: null,
+      };
+
+      // a Timestamp-Like object (we lose the instance after serialization)
+      const end_date = JSON.parse(
+        JSON.stringify(admin.firestore.Timestamp.now())
+      );
+
+      const event: FirestoreDocumentChangeEvent = {
+        timestamp: "",
+        operation: ChangeType.CREATE,
+        documentName: "",
+        eventId: "",
+        documentId: "",
+        data: { end_date },
+      };
+
+      const partitioning = new Partitioning(config, table);
+      const value = partitioning.getPartitionValue(event);
+
+      expect(value.end_date).toBeDefined();
+    });
+
     test("returns an empty object if timePartitioningFirestoreField has not been provided", async () => {
       const config: FirestoreBigQueryEventHistoryTrackerConfig = {
         datasetId: "",
@@ -372,5 +477,55 @@ describe("processing partitions on a new table", () => {
     await deleteTable({
       datasetId,
     });
+  });
+});
+
+describe("updateTableMetadata", () => {
+  test("updates the table metadata with the timestamp field", async () => {
+    const config: FirestoreBigQueryEventHistoryTrackerConfig = {
+      datasetId: "",
+      tableId: "",
+      datasetLocation: "",
+      timePartitioning: "MONTH",
+      timePartitioningField: "timestamp",
+      timePartitioningFieldType: undefined,
+      timePartitioningFirestoreField: undefined,
+      transformFunction: "",
+      clustering: [],
+      bqProjectId: null,
+    };
+    const options = {};
+
+    const partitioning = new Partitioning(config, table);
+
+    await partitioning.updateTableMetadata(options);
+
+    expect(options).toEqual({
+      timePartitioning: {
+        field: "timestamp",
+        type: "MONTH",
+      },
+    });
+  });
+  test("Should not update if there is a custom option with the timestamp option", async () => {
+    const config: FirestoreBigQueryEventHistoryTrackerConfig = {
+      datasetId: "",
+      tableId: "",
+      datasetLocation: "",
+      timePartitioning: "MONTH",
+      timePartitioningField: "timestamp",
+      timePartitioningFieldType: "DATETIME",
+      timePartitioningFirestoreField: undefined,
+      transformFunction: "",
+      clustering: [],
+      bqProjectId: null,
+    };
+    const options = {};
+
+    const partitioning = new Partitioning(config, table);
+
+    await partitioning.updateTableMetadata(options);
+
+    expect(options).toEqual({});
   });
 });