+++ firestore-shorten-urls-bitly/CHANGELOG.md
@@ -1,3 +1,11 @@
+## Version 0.1.14
+
+feature - add optional custom events
+
+## Version 0.1.13
+
+build - updated depenencies
+
 ## Version 0.1.12
 
 feature - bump to node 18